+++ admin/user_identification_information/views.py
@@ -1,8 +1,6 @@
 import csv
 from operator import itemgetter
 
-import numpy as np
-from django.db import connection
 from django.db.models import Q
 from django.http import Http404
 from django.http import HttpResponse
@@ -10,77 +8,13 @@ from django.views.generic import ListView
 
 from admin.base.views import GuidView
 from admin.rdm.utils import RdmPermissionMixin
+from admin.user_identification_information.utils import (
+    custom_size_abbreviation,
+    get_list_extend_storage
+)
 from api.base import settings as api_settings
-from osf.models import ExternalAccount
 from osf.models import OSFUser, UserQuota, Email
 from website.util import quota
-
-
-def custom_size_abbreviation(size, abbr, *kwargs):
-    if abbr == 'B':
-        return (size / api_settings.BASE_FOR_METRIC_PREFIX, 'KB')
-    return size, abbr
-
-
-def get_list_extend_storage():
-    values = ExternalAccount.objects.values_list('provider', 'provider_name')
-    get_provider, get_provider_name = map(list, zip(*values))
-    dict_users_list = {}
-    storage_branch_name = None
-    cursor = connection.cursor()
-
-    for j in range(len(get_provider)):
-        provider_value = get_provider[j]
-        get_provider_name_value = get_provider_name[j]
-        if any(s in provider_value.lower() for s in
-               ('s3', 's3compat', 's3compatb3', 'azureblobstorage', 'box',
-                'figshare', 'onedrivebusiness', 'swift')):
-            storage_branch_name = 'folder_name'
-        elif any(s in provider_value.lower() for s in ('bitbucket', 'github',
-                                                       'gitlab')):
-            storage_branch_name = 'repo'
-        elif any(s in provider_value.lower() for s in ('googledrive',
-                                                       'onedrive', 'iqbrims')):
-            storage_branch_name = 'folder_path'
-        elif any(s in provider_value.lower() for s in ('dropbox')):
-            storage_branch_name = 'folder'
-        elif any(s in provider_value.lower() for s in ('weko')):
-            storage_branch_name = 'index_title'
-        elif any(s in provider_value.lower() for s in ('mendeley', 'zotero')):
-            storage_branch_name = 'list_id'
-        elif any(s in provider_value.lower() for s in ('owncloud')):
-            storage_branch_name = 'folder_id'
-        elif any(s in provider_value.lower() for s in ('dataverse')):
-            storage_branch_name = 'dataverse'
-
-        cursor.execute(
-            """
-            select addons_%s_nodesettings.%s, addons_%s_usersettings.owner_id as user_id
-            from addons_%s_usersettings inner join addons_%s_nodesettings
-            on addons_%s_nodesettings.user_settings_id = addons_%s_usersettings.id
-            where addons_%s_usersettings.id in(
-                select addons_%s_usersettings.id from osf_osfuser inner join addons_%s_usersettings
-                on osf_osfuser.id = addons_%s_usersettings.owner_id)
-            """ % (
-                provider_value, storage_branch_name, provider_value, provider_value, provider_value, provider_value,
-                provider_value, provider_value, provider_value, provider_value, provider_value)
-        )
-        result = np.asarray(cursor.fetchall())
-        list_users_provider = result[:, 0]
-        list_users_id = result[:, 1]
-
-        for i in range(len(list_users_id)):
-            if list_users_id[i] not in dict_users_list:
-                dict_users_list[list_users_id[i]] = [
-                    list_users_provider[i] + '/' +
-                    get_provider_name_value if list_users_provider[i] is not None else '/' + get_provider_name_value]
-            else:
-                current_val = dict_users_list.get(list_users_id[i])
-                current_val.append(
-                    list_users_provider[i] + '/' +
-                    get_provider_name_value if list_users_provider[i] is not None else '/' + get_provider_name_value)
-                dict_users_list[list_users_id[i]] = current_val
-        return list_users_id, dict_users_list
 
 
 class UserIdentificationInformation(ListView):
@@ -152,9 +86,7 @@ class UserIdentificationList(RdmPermissionMixin, UserIdentificationInformation):
     raise_exception = True
     paginate_by = 20
 
-    def get_userlist(self):
-        if self.is_admin:
-            raise Http404('Page not found')
+    def user_list(self):
         guid = self.request.GET.get('guid')
         name = self.request.GET.get('fullname')
         email = self.request.GET.get('username')
@@ -195,15 +127,18 @@ class UserIdentificationList(RdmPermissionMixin, UserIdentificationInformation):
         else:
             return []
 
+    def get_userlist(self):
+        if self.is_admin:
+            raise Http404('Page not found')
+        return self.user_list()
+
 
 class UserIdentificationDetails(RdmPermissionMixin, GuidView):
     template_name = 'user_identification_information/user_identification_details.html'
     context_object_name = 'user_details'
     raise_exception = True
 
-    def get_object(self):
-        if self.is_admin:
-            raise Http404('Page not found')
+    def user_details(self):
         user_details = OSFUser.load(self.kwargs.get('guid'))
         user_id = int(user_details.id)
         max_quota, used_quota = quota.get_quota_info(user_details, UserQuota.NII_STORAGE)
@@ -241,6 +176,11 @@ class UserIdentificationDetails(RdmPermissionMixin, GuidView):
             'remaining_abbr': remaining_abbr[1],
             'extended_storage': extend_storage,
         }
+
+    def get_object(self):
+        if self.is_admin:
+            raise Http404('Page not found')
+        return self.user_details()
 
 
 class ExportFileCSV(RdmPermissionMixin, UserIdentificationInformation):
