+++ admin/user_identification_information_admin/urls.py
@@ -5,8 +5,8 @@ from . import views
 app_name = 'admin'
 
 urlpatterns = [
-    url(r'^$', views.UserIdentificationList.as_view(), name='index'),
-    url(r'^csvexport/$', views.ExportFileCSV.as_view(), name='csvexport'),
-    url(r'^(?P<guid>[a-z0-9]+)/$', views.UserIdentificationDetails.as_view(), name='details'),
+    url(r'^$', views.UserIdentificationListAdminView.as_view(), name='index'),
+    url(r'^csvexport/$', views.ExportFileCSVAdminView.as_view(), name='csvexport'),
+    url(r'^(?P<guid>[a-z0-9]+)/$', views.UserIdentificationDetailsAdminView.as_view(), name='details'),
 
 ]
